// source/api/v1/shared/exceptions/UserExceptions.ts
import { Exception } from "../utils/typing/Exception";


const USER_ALREADY_EXISTS: Exception = {
    statusCode: 400,
    message: "User already exists"
}


const USER_NOT_FOUND: Exception = {
    statusCode: 404,
    message: "User not found"
}

const USER_NOT_AUTHORIZED: Exception = {
    statusCode: 401,
    message:
        "Authorization token niether is not valid or is not generated by server"
}

export {
    USER_ALREADY_EXISTS,
    USER_NOT_AUTHORIZED,
    USER_NOT_FOUND
}